Struct moore_svlog_hir::HierarchyBody

pub struct HierarchyBody {
    pub procs: Vec<Procedure>,
    pub nets: Vec<NetDecl>,
    pub vars: Vec<VarDecl>,
    pub assigns: Vec<ContAssign>,
    pub params: Vec<ParamDecl>,
    pub insts: Vec<Inst>,
    pub genreg: Vec<HierarchyBody>,
    pub genvars: Vec<GenvarDecl>,
    pub genfors: Vec<GenerateFor>,
    pub genifs: Vec<GenerateIf>,
    pub gencases: Vec<GenerateCase>,
    pub classes: Vec<ClassDecl>,
    pub subroutines: Vec<SubroutineDecl>,
    pub asserts: Vec<Assertion>,
    pub typedefs: Vec<Typedef>,
}

A hierarchy body represents the contents of a module, interface, or package. Generate regions and nested modules introduce additional bodies. The point of hierarchy bodies is to take a level of the design hierarchy and group all declarations by type, rather than having them in a single array in declaration order.

Fields
procs: Vec<Procedure> nets: Vec<NetDecl> vars: Vec<VarDecl> assigns: Vec<ContAssign> params: Vec<ParamDecl> insts: Vec<Inst> genreg: Vec<HierarchyBody> genvars: Vec<GenvarDecl> genfors: Vec<GenerateFor> genifs: Vec<GenerateIf> gencases: Vec<GenerateCase> classes: Vec<ClassDecl> subroutines: Vec<SubroutineDecl> asserts: Vec<Assertion> typedefs: Vec<Typedef>