Struct moore_svlog_hir::Span

pub struct Span {
    pub source: Source,
    pub begin: usize,
    pub end: usize,
}

A span of locations within a source file, expressed as a half-open interval of bytes [begin,end).

Fields
source: Source begin: usize end: usize
Methods

impl Span

fn new(source: Source, begin: usize, end: usize) -> Span

Create a new span from two byte offsets.

fn union<S>(a: S, b: S) -> Span where
S: Into<Span>,

Create a new span that covers two spans, i.e. represents the smallest possible span that fully contains both input spans a and b.

fn expand<S>(&mut self, other: S) -> &mut Span where
S: Into<Span>,

Modify this range to also cover the entirety of the other range. The other range must lie in the same source as self.

fn begin(&self) -> Location

Return the location just before the first character in this span.

fn end(&self) -> Location

Return the location just after the last character in this span.

fn extract(&self) -> String

Copy the portion of the source file in this span into an owned string.

fn iter(
self,
content: &'a Rc<SourceContent + 'static>
) -> Box<DoubleEndedIterator<Item = (usize, char)> + 'a>

Obtain an iterator over the extract of the source file describe by this span.